src/urbvan/requirements/base.txt
Django==2.1.2
django-rest-framework==0.1.0
djangorestframework==3.8.2
factory-boy==2.11.1
marshmallow==2.15.6
